chore: extract primitives into their own crate
